+++ packages/ts-plugin/src/language-service/proxy.ts
@@ -1,19 +1,22 @@
+import type { CMKConfig } from '@css-modules-kit/core';
 import { createExportBuilder, type MatchesPattern, type Resolver } from '@css-modules-kit/core';
 import type { Language } from '@volar/language-core';
 import type ts from 'typescript';
 import { CMK_DATA_KEY, isCSSModuleScript } from '../language-plugin.js';
 import { getCodeFixesAtPosition } from './feature/code-fix.js';
-import { getCompletionsAtPosition } from './feature/completion.js';
+import { getCompletionEntryDetails, getCompletionsAtPosition } from './feature/completion.js';
 import { getApplicableRefactors, getEditsForRefactor } from './feature/refactor.js';
 import { getSemanticDiagnostics } from './feature/semantic-diagnostic.js';
 import { getSyntacticDiagnostics } from './feature/syntactic-diagnostic.js';
 
+// eslint-disable-next-line max-params
 export function proxyLanguageService(
   language: Language<string>,
   languageService: ts.LanguageService,
   project: ts.server.Project,
   resolver: Resolver,
   matchesPattern: MatchesPattern,
+  config: CMKConfig,
 ): ts.LanguageService {
   const proxy: ts.LanguageService = Object.create(null);
 
@@ -44,8 +47,9 @@ export function proxyLanguageService(
   );
   proxy.getApplicableRefactors = getApplicableRefactors(languageService, project);
   proxy.getEditsForRefactor = getEditsForRefactor(languageService);
-  proxy.getCompletionsAtPosition = getCompletionsAtPosition(languageService);
-  proxy.getCodeFixesAtPosition = getCodeFixesAtPosition(language, languageService, project);
+  proxy.getCompletionsAtPosition = getCompletionsAtPosition(languageService, config);
+  proxy.getCompletionEntryDetails = getCompletionEntryDetails(languageService, resolver, config);
+  proxy.getCodeFixesAtPosition = getCodeFixesAtPosition(language, languageService, project, resolver, config);
 
   return proxy;
 }