+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @css-modules-kit/core
 
+## 0.3.0
+
+### Minor Changes
+
+- 91c21eb: feat: support Node.js >=20.19.0
+- fb0563d: feat: support `prioritizeNamedImports` option
+- 9b40191: feat: support `namedExports` option
+
 ## 0.2.0
 
 ### Minor Changes