+++ packages/ts-plugin/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @css-modules-kit/ts-plugin
 
+## 0.2.0
+
+### Minor Changes
+
+- 91c21eb: feat: support Node.js >=20.19.0
+- fb0563d: feat: support `prioritizeNamedImports` option
+- 9b40191: feat: support `namedExports` option
+
+### Patch Changes
+
+- 3ec5b22: fix: exclude files in `generated/` from completion candidates
+- Updated dependencies [91c21eb]
+- Updated dependencies [fb0563d]
+- Updated dependencies [9b40191]
+  - @css-modules-kit/core@0.3.0
+
 ## 0.1.1
 
 ### Patch Changes